requirements.txt
accelerate==1.12.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
clip-anytorch==2.6.0
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
diffusers==0.36.0
einops==0.8.1
filelock==3.20.3
fonttools==4.61.1
fsspec==2026.1.0
ftfy==6.3.1
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
ImageIO==2.37.2
importlib_metadata==8.7.1
Jinja2==3.1.6
kiwisolver==1.4.9
lazy_loader==0.4
lpips==0.1.4
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
networkx==3.6.1
numpy==2.2.6
omegaconf==2.3.0
open_clip_torch==3.2.0
opencv-python==4.12.0.88
packaging==25.0
pillow==12.1.0
psutil==7.2.1
pyparsing==3.3.1
python-dateutil==2.9.0.post0
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
safetensors==0.7.0
scikit-image==0.26.0
scipy==1.17.0
shellingham==1.5.4
six==1.17.0
sympy==1.14.0
tifffile==2026.1.14
timm==1.0.24
tokenizers==0.22.2
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
transformers==4.57.6
typer-slim==0.21.1
typing_extensions==4.15.0
urllib3==2.6.3
wcwidth==0.2.14
zipp==3.23.0
